console,account: Move online status store logic to global util
