+++ pkg/webui/lib/hooks/use-request.js
@@ -16,8 +16,7 @@ import { useState, useEffect, useRef } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 
 import attachPromise from '@ttn-lw/lib/store/actions/attach-promise'
-
-import { selectIsOnlineStatus } from '@console/store/selectors/status'
+import { selectIsOnlineStatus } from '@ttn-lw/lib/store/selectors/status'
 
 const useRequest = requestAction => {
   const dispatch = useDispatch()
