+++ pkg/webui/lib/store/actions/status.js
